Add interface for emitter sampler Move shared RayTracer methods to utility class
