+++ chunky/src/java/se/llbit/chunky/renderer/scene/PathTracer.java
@@ -16,48 +16,54 @@
  */
 package se.llbit.chunky.renderer.scene;
 
+import java.util.Random;
 import org.apache.commons.math3.util.FastMath;
 import se.llbit.chunky.block.Air;
 import se.llbit.chunky.block.Water;
 import se.llbit.chunky.model.WaterModel;
 import se.llbit.chunky.renderer.EmitterSamplingStrategy;
 import se.llbit.chunky.renderer.WorkerState;
+import se.llbit.chunky.renderer.scene.EmitterSamplerFactory.EmitterSampler;
 import se.llbit.chunky.world.Material;
-import se.llbit.math.*;
-
-import java.util.Random;
+import se.llbit.math.QuickMath;
+import se.llbit.math.Ray;
+import se.llbit.math.Vector3;
+import se.llbit.math.Vector4;
 
 /**
- * Static methods for path tracing.
+ * Basic path tracer.
  *
  * @author Jesper Öqvist <jesper@llbit.se>
  */
 public class PathTracer implements RayTracer {
 
-  /** Extinction factor for fog rendering. */
+  /**
+   * Extinction factor for fog rendering.
+   */
   private static final double EXTINCTION_FACTOR = 0.04;
 
   /**
    * Path trace the ray.
    */
-  @Override public void trace(Scene scene, WorkerState state) {
+  @Override
+  public void trace(Scene scene, WorkerState state) {
     Ray ray = state.ray;
     if (scene.isInWater(ray)) {
       ray.setCurrentMaterial(Water.INSTANCE);
     } else {
       ray.setCurrentMaterial(Air.INSTANCE);
     }
-    pathTrace(scene, ray, state, 1, true);
+    pathTrace(scene, ray, state, 1, true, scene.getEmitterSampler());
   }
 
   /**
    * Path trace the ray in this scene.
    *
-   * @param firstReflection {@code true} if the ray has not yet hit the first
-   * diffuse or specular reflection
+   * @param firstReflection {@code true} if the ray has not yet hit the first diffuse or specular
+   *                        reflection
    */
-  public static boolean pathTrace(Scene scene, Ray ray, WorkerState state, int addEmitted,
-      boolean firstReflection) {
+  private static boolean pathTrace(Scene scene, Ray ray, WorkerState state, int addEmitted,
+      boolean firstReflection, EmitterSampler emitterSampler) {
 
     boolean hit = false;
     Random random = state.random;
@@ -67,7 +73,7 @@ public class PathTracer implements RayTracer {
 
     while (true) {
 
-      if (!PreviewRayTracer.nextIntersection(scene, ray)) {
+      if (!RayTracers.nextIntersection(scene, ray)) {
         if (ray.getPrevMaterial().isWater()) {
           ray.color.set(0, 0, 0, 1);
           hit = true;
@@ -166,7 +172,9 @@ public class PathTracer implements RayTracer {
 
             Vector4 indirectEmitterColor = new Vector4(0, 0, 0, 0);
 
-            if (scene.emittersEnabled && (!scene.isPreventNormalEmitterWithSampling() || scene.getEmitterSamplingStrategy() == EmitterSamplingStrategy.NONE || ray.depth == 0) && currentMat.emittance > Ray.EPSILON) {
+            if (scene.emittersEnabled && (!scene.isPreventNormalEmitterWithSampling()
+                || scene.getEmitterSamplingStrategy() == EmitterSamplingStrategy.NONE
+                || ray.depth == 0) && currentMat.emittance > Ray.EPSILON) {
 
               emittance = addEmitted;
               ray.emittance.x = ray.color.x * ray.color.x *
@@ -176,19 +184,11 @@ public class PathTracer implements RayTracer {
               ray.emittance.z = ray.color.z * ray.color.z *
                   currentMat.emittance * scene.emitterIntensity;
               hit = true;
-            } else if(scene.emittersEnabled && scene.emitterSamplingStrategy != EmitterSamplingStrategy.NONE && scene.getEmitterGrid() != null) {
-              // Sample emitter
-              boolean sampleOne = scene.emitterSamplingStrategy == EmitterSamplingStrategy.ONE;
-              if(sampleOne) {
-                Grid.EmitterPosition pos = scene.getEmitterGrid().sampleEmitterPosition((int) ray.o.x, (int) ray.o.y, (int) ray.o.z, random);
-                if(pos != null) {
-                  indirectEmitterColor = sampleEmitter(scene, ray, pos,  random);
-                }
-              } else {
-                for(Grid.EmitterPosition pos : scene.getEmitterGrid().getEmitterPositions((int) ray.o.x, (int) ray.o.y, (int) ray.o.z)) {
-                  indirectEmitterColor.scaleAdd(1, sampleEmitter(scene, ray, pos, random));
-                }
-              }
+            } else if (scene.emittersEnabled
+                && scene.emitterSamplingStrategy != EmitterSamplingStrategy.NONE
+                && scene.getEmitterGrid() != null) {
+
+              indirectEmitterColor = emitterSampler.sample(scene, ray, random);
             }
 
             if (scene.sunEnabled) {
@@ -231,7 +231,8 @@ public class PathTracer implements RayTracer {
                     reflected.color.y + reflected.emittance.y) + (indirectEmitterColor.y));
                 ray.color.z = ray.color.z * (emittance + directLightB * scene.sun.emittance.z + (
                     reflected.color.z + reflected.emittance.z) + (indirectEmitterColor.z));
-              } else if(indirectEmitterColor.x > Ray.EPSILON || indirectEmitterColor.y > Ray.EPSILON || indirectEmitterColor.z > Ray.EPSILON) {
+              } else if (indirectEmitterColor.x > Ray.EPSILON
+                  || indirectEmitterColor.y > Ray.EPSILON || indirectEmitterColor.z > Ray.EPSILON) {
                 hit = true;
                 ray.color.x *= indirectEmitterColor.x;
                 ray.color.y *= indirectEmitterColor.y;
@@ -244,12 +245,16 @@ public class PathTracer implements RayTracer {
               hit = pathTrace(scene, reflected, state, 0, false) || hit;
               if (hit) {
                 ray.color.x =
-                    ray.color.x * (emittance + (reflected.color.x + reflected.emittance.x) + (indirectEmitterColor.x));
+                    ray.color.x * (emittance + (reflected.color.x + reflected.emittance.x)
+                        + (indirectEmitterColor.x));
                 ray.color.y =
-                    ray.color.y * (emittance + (reflected.color.y + reflected.emittance.y) + (indirectEmitterColor.y));
+                    ray.color.y * (emittance + (reflected.color.y + reflected.emittance.y)
+                        + (indirectEmitterColor.y));
                 ray.color.z =
-                    ray.color.z * (emittance + (reflected.color.z + reflected.emittance.z) + (indirectEmitterColor.z));
-              } else if(indirectEmitterColor.x > Ray.EPSILON || indirectEmitterColor.y > Ray.EPSILON || indirectEmitterColor.z > Ray.EPSILON) {
+                    ray.color.z * (emittance + (reflected.color.z + reflected.emittance.z)
+                        + (indirectEmitterColor.z));
+              } else if (indirectEmitterColor.x > Ray.EPSILON
+                  || indirectEmitterColor.y > Ray.EPSILON || indirectEmitterColor.z > Ray.EPSILON) {
                 hit = true;
                 ray.color.x *= indirectEmitterColor.x;
                 ray.color.y *= indirectEmitterColor.y;
@@ -356,7 +361,7 @@ public class PathTracer implements RayTracer {
 
       if (hit && prevMat.isWater()) {
         // Render water fog effect.
-        if(scene.waterVisibility == 0) {
+        if (scene.waterVisibility == 0) {
           ray.color.scale(0.);
         } else {
           double a = ray.distance / scene.waterVisibility;
@@ -423,58 +428,9 @@ public class PathTracer implements RayTracer {
   }
 
   /**
-   * Cast a shadow ray from the intersection point (given by ray) to the emitter
-   * at position pos. Returns the contribution of this emitter (0 if the emitter is occluded)
-   * @param scene The scene being rendered
-   * @param ray The ray that generated the intersection
-   * @param pos The position of the emitter to sample
-   * @param random RNG
-   * @return The contribution of the emitter
-   */
-  private static Vector4 sampleEmitter(Scene scene, Ray ray, Grid.EmitterPosition pos, Random random) {
-    Vector4 indirectEmitterColor = new Vector4();
-    Ray emitterRay = new Ray();
-    emitterRay.set(ray);
-    // TODO Sampling a random point on the model would be better than using a random point in the middle of the cube
-    Vector3 target = new Vector3(pos.x + (random.nextDouble() - 0.5) * pos.radius, pos.y + (random.nextDouble() - 0.5) * pos.radius, pos.z  + (random.nextDouble() - 0.5) * pos.radius);
-    emitterRay.d.set(target);
-    emitterRay.d.sub(emitterRay.o);
-    double distance = emitterRay.d.length();
-    emitterRay.d.normalize();
-    double indirectEmitterCoef = emitterRay.d.dot(emitterRay.n);
-    if(indirectEmitterCoef > 0) {
-      // Here We need to invert the material.
-      // The fact that the dot product is > 0 guarantees that the ray is going away from the surface
-      // it just met. This means the ray is going from the block just hit to the previous material (usually air or water)
-      // TODO If/when normal mapping is implemented, indirectEmitterCoef will be computed with the mapped normal
-      //      but the dot product with the original geometry normal will still need to be computed
-      //      to ensure the emitterRay isn't going through the geometry
-      Material prev = emitterRay.getPrevMaterial();
-      int prevData = emitterRay.getPrevData();
-      emitterRay.setPrevMaterial(emitterRay.getCurrentMaterial(), emitterRay.getCurrentData());
-      emitterRay.setCurrentMaterial(prev, prevData);
-      emitterRay.emittance.set(0, 0, 0);
-      emitterRay.o.scaleAdd(Ray.EPSILON, emitterRay.d);
-      PreviewRayTracer.nextIntersection(scene, emitterRay);
-      if(emitterRay.getCurrentMaterial().emittance > Ray.EPSILON) {
-        indirectEmitterColor.set(emitterRay.color);
-        indirectEmitterColor.scale(emitterRay.getCurrentMaterial().emittance);
-        // TODO Take fog into account
-        indirectEmitterCoef *= scene.emitterIntensity;
-        // Dont know if really realistic but offer better convergence and is better artistically
-        indirectEmitterCoef /= Math.max(distance * distance, 1);
-      }
-    } else {
-      indirectEmitterCoef = 0;
-    }
-    indirectEmitterColor.scale(indirectEmitterCoef);
-    return indirectEmitterColor;
-  }
-
-  /**
    * Calculate direct lighting attenuation.
    */
-  public static void getDirectLightAttenuation(Scene scene, Ray ray, WorkerState state) {
+  private static void getDirectLightAttenuation(Scene scene, Ray ray, WorkerState state) {
 
     Vector4 attenuation = state.attenuation;
     attenuation.x = 1;
@@ -483,7 +439,7 @@ public class PathTracer implements RayTracer {
     attenuation.w = 1;
     while (attenuation.w > 0) {
       ray.o.scaleAdd(Ray.OFFSET, ray.d);
-      if (!PreviewRayTracer.nextIntersection(scene, ray)) {
+      if (!RayTracers.nextIntersection(scene, ray)) {
         break;
       }
       double mult = 1 - ray.color.w;
@@ -492,7 +448,7 @@ public class PathTracer implements RayTracer {
       attenuation.z *= ray.color.z * ray.color.w + mult;
       attenuation.w *= mult;
       if (ray.getPrevMaterial().isWater()) {
-        if(scene.waterVisibility == 0) {
+        if (scene.waterVisibility == 0) {
           attenuation.w = 0;
         } else {
           double a = ray.distance / scene.waterVisibility;
